Refactored (#7 Delivery zone planning)

+++ Andrei15193.Edesia/Controllers/ProductController.cs
@@ -55,7 +55,9 @@ namespace Andrei15193.Edesia.Controllers
 		[HttpGet]
 		public ActionResult Remove(string productName)
 		{
-			_productRepository.RemoveProduct(productName);
+			if (productName != null)
+				_productRepository.RemoveProduct(productName);
+
 			return RedirectToAction("Default", "Product");
 		}
